+++ modules/plugin/jdbc/jdbc-oracle/src/test/java/org/geotools/data/oracle/OracleConnectionLifecycleOnlineTest.java
@@ -2,7 +2,7 @@
  *    GeoTools - The Open Source Java GIS Toolkit
  *    http://geotools.org
  *
- *    (C) 2002-2008, Open Source Geospatial Foundation (OSGeo)
+ *    (C) 2002-2015, Open Source Geospatial Foundation (OSGeo)
  *
  *    This library is free software; you can redistribute it and/or
  *    modify it under the terms of the GNU Lesser General Public
@@ -16,7 +16,14 @@
  */
 package org.geotools.data.oracle;
 
+import java.util.HashMap;
+import java.util.Properties;
+
+import org.geotools.data.DataStore;
+import org.geotools.data.DataStoreFinder;
 import org.geotools.jdbc.JDBCConnectionLifecycleOnlineTest;
+import org.geotools.jdbc.JDBCDataStore;
+import org.geotools.jdbc.JDBCDataStoreFactory;
 import org.geotools.jdbc.JDBCTestSetup;
 
 public class OracleConnectionLifecycleOnlineTest extends JDBCConnectionLifecycleOnlineTest {
@@ -26,4 +33,23 @@ public class OracleConnectionLifecycleOnlineTest extends JDBCConnectionLifecycle
         return new OracleTestSetup();
     }
 
+    public void testLifeCycleDoubleUnwrap() {
+        try {
+            // Use startup SQL when connecting so the connection is
+            // doubly wrapped (adding LifeCycleConnection).
+            // That tests ability of OracleDialect to unwrap properly.
+            Properties addStartupSql = (Properties) fixture.clone();
+            addStartupSql.setProperty(JDBCDataStoreFactory.SQL_ON_BORROW.key,
+                    "select sysdate from dual");
+            HashMap params = createDataStoreFactoryParams();
+            params.putAll(addStartupSql);
+            DataStore withWrap = (JDBCDataStore) DataStoreFinder.getDataStore(params);
+            if (withWrap == null) {
+                throw new RuntimeException("Failed to create DataStore with startup sql");
+            }
+            withWrap.dispose();
+        } catch (Exception e) {
+            throw new RuntimeException("Connection unwrap test failed", e);
+        }
+    }
 }
